refactor(tanstack-query): move operation key types next to generateOperationKey
