+++ packages/tanstack-query/src/key.ts
@@ -1,4 +1,24 @@
-import type { OperationKey, OperationKeyOptions, OperationType } from './types'
+import type { PartialDeep } from '@orpc/shared'
+import type { SerializableStreamedQueryOptions } from './stream-query'
+import type { OperationType } from './types'
+
+export interface OperationKeyPrefixOptions {
+  /**
+   * Prepended as the first element of the key when present.
+   * Use this to avoid key conflicts when multiple router utils share the same client.
+   */
+  prefix?: string
+}
+
+export type OperationKeyOptions<TType extends OperationType, TInput> = OperationKeyPrefixOptions & {
+  type?: TType
+  input?: TType extends 'mutation' ? never : PartialDeep<TInput>
+  fnOptions?: TType extends 'streamed' ? SerializableStreamedQueryOptions : never
+}
+
+export type OperationKey<TType extends OperationType, TInput>
+  = | [path: string[], options: Omit<OperationKeyOptions<TType, TInput>, 'prefix'>]
+    | [prefix: string, path: string[], options: Omit<OperationKeyOptions<TType, TInput>, 'prefix'>]
 
 export function generateOperationKey<TType extends OperationType, TInput>(
   path: string[],
